+++ libandroid-navigation/src/main/java/com/mapbox/services/android/navigation/v5/internal/accounts/TokenGenerator.kt
@@ -1,4 +1,4 @@
-package com.mapbox.services.android.navigation.v5.accounts
+package com.mapbox.services.android.navigation.v5.internal.accounts
 
 internal interface TokenGenerator {
     fun obtainSkuToken(): String
