+++ libandroid-navigation/src/test/java/com/mapbox/services/android/navigation/v5/navigation/OfflineTilesTest.kt
@@ -1,5 +1,6 @@
 package com.mapbox.services.android.navigation.v5.navigation
 
+import android.content.Context
 import com.mapbox.api.routetiles.v1.MapboxRouteTiles
 import org.junit.Test
 import org.mockito.Mockito.mock
@@ -9,8 +10,8 @@ class OfflineTilesTest {
     @Test(expected = IllegalStateException::class)
     fun checkOfflineTilesBuildIfNoVersionThrowException() {
         val mockMapboxRouteTilesBuilder = mock(MapboxRouteTiles.Builder::class.java)
-
-        val builder = OfflineTiles.builder(mockMapboxRouteTilesBuilder).accessToken("")
+        val mockContext = mock(Context::class.java)
+        val builder = OfflineTiles.builder(mockMapboxRouteTilesBuilder, mockContext).accessToken("")
         builder.build()
     }
 }
